+++ lib/flutter_iconify.dart
@@ -2,3 +2,4 @@ library flutter_iconify;
 
 export 'src/iconify_lighter_icons.dart';
 export 'src/iconify_material_icons.dart';
+export 'src/iconify_bolder_icons.dart';