+++ lib/src/iconify_bolder_icons.dart
@@ -1,0 +1,25 @@
+// ignore_for_file: constant_identifier_names
+
+import 'package:flutter/widgets.dart';
+
+class IconifyBolder {
+  IconifyBolder._();
+
+  static const _kFontFam = 'IconifyBolder';
+  static const String _kFontPkg = "flutter_iconify";
+
+  static const IconData home_filled =
+      IconData(0xe800, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData home =
+      IconData(0xe801, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData account =
+      IconData(0xe802, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData bonds =
+      IconData(0xe803, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData ad_blocker =
+      IconData(0xe804, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData portfolio =
+      IconData(0xe805, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData circled_menu =
+      IconData(0xe806, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+}